polish: Fix lint issues and improve code quality for Claude's integration
- Fix S324 (insecure hash): Replace md5 with sha256 in automl optimizer
- Fix RUF006: Store asyncio.create_task reference in raft_consensus
- Fix F401: Clean up unused qiskit imports in quantum_computing/circuits
- Add per-file ruff ignores for domain-specific patterns:
  - E741 for harmonics (standard 'l' notation for spherical harmonic degree)
  - SIM113 for advanced detectors (intentional batch counting)
  - S310 for financial/weather stubs (trusted HTTPS URLs)
- Fix test_train_model_with_data: Provide 2 samples and expect VALIDATION phase
- Apply ruff auto-fixes for TC003, RUF022, RUF046, F841 issues

+++ src/omni_mercury_engine/automl/search_space.py
@@ -13,6 +13,7 @@ from typing import Any
 
 import numpy as np
 
+
 logger = logging.getLogger(__name__)
 
 
@@ -146,7 +147,7 @@ class IntUniformParameter(HyperParameter):
 
     def from_normalized(self, normalized: float) -> int:
         """Convert from [0, 1]."""
-        return int(round(self._low + normalized * (self._high - self._low)))
+        return round(self._low + normalized * (self._high - self._low))
 
     def get_bounds(self) -> tuple[int, int]:
         """Get bounds."""
@@ -179,7 +180,7 @@ class CategoricalParameter(HyperParameter):
 
     def from_normalized(self, normalized: float) -> Any:
         """Convert from [0, 1]."""
-        idx = int(round(normalized * (len(self._choices) - 1)))
+        idx = round(normalized * (len(self._choices) - 1))
         return self._choices[idx]
 
     def get_bounds(self) -> tuple[int, int]:
